feat: Add roster.h/.cpp code, new test in main.cpp
initial structure of roster.h and roster.cpp added.

new test code required the addition of a constructor declaration
and implementation in the student.h and student.cpp files, an explicit
default constructor was defined as well in order to allow original test
code to function (once a parameterized constructor is defined,
the compiler no longer provides this default constructor automatically
unless one explicitly defines it).